:recycle: 게임 제어 컨트롤러 통합
